+++ knex/migrations/20210124181241_create_transactions_table.js
@@ -8,12 +8,12 @@ exports.up = function(knex) {
             tbl.string(table.field.recipientAddress, 50).nullable().index();
             tbl.bigInteger(table.field.amount).nullable();
             tbl.bigInteger(table.field.fee).notNullable();
-            tbl.integer(table.field.timestamp).notNullable().index();
+            tbl.bigInteger(table.field.timestamp).notNullable().index();
             tbl.text(table.field.message).notNullable();
             tbl.string(table.field.senderAddress, 50).notNullable().index();
             tbl.string(table.field.sigPublicKey, 64).nullable();
             tbl.string(table.field.nextSigPublicKey, 64).nullable();
-            tbl.integer(table.field.nextSigKeyIndex).nullable();
+            tbl.bigInteger(table.field.nextSigKeyIndex).nullable();
             tbl.string(table.field.senderSignatureHash, 44).nullable();
             tbl.text(table.field.signatures).nullable();
             tbl.string(table.field.blockId, 40).notNullable().index();
@@ -22,13 +22,13 @@ exports.up = function(knex) {
             tbl.string(table.field.delegateAddress, 50).nullable();
             tbl.string(table.field.newForgingPublicKey, 64).nullable();
             tbl.string(table.field.newNextForgingPublicKey, 64).nullable();
-            tbl.integer(table.field.newNextForgingKeyIndex).nullable();
+            tbl.bigInteger(table.field.newNextForgingKeyIndex).nullable();
             tbl.string(table.field.newMultisigPublicKey, 64).nullable();
             tbl.string(table.field.newNextMultisigPublicKey, 64).nullable();
-            tbl.integer(table.field.newNextMultisigKeyIndex).nullable();
+            tbl.bigInteger(table.field.newNextMultisigKeyIndex).nullable();
             tbl.string(table.field.newSigPublicKey, 64).nullable();
             tbl.string(table.field.newNextSigPublicKey, 64).nullable();
-            tbl.integer(table.field.newNextSigKeyIndex).nullable();
+            tbl.bigInteger(table.field.newNextSigKeyIndex).nullable();
             // todo : need to check if specific type can be kept as text/json ( JSON serialized)
             tbl.specificType(table.field.memberAddresses,'text ARRAY').nullable();
             tbl.integer(table.field.requiredSignatureCount).nullable();
